requirements.txt
streamlit
streamlit_space
streamlit_shadcn_ui
python-dotenv
altair
openai